Created better functional movement system
