+++ .vscode/launch.json
@@ -1,0 +1,22 @@
+{
+    // Use IntelliSense to find out which attributes exist for C# debugging
+    // Use hover for the description of the existing attributes
+    // For further information visit https://github.com/OmniSharp/omnisharp-vscode/blob/master/debugger-launchjson.md
+    "version": "0.2.0",
+    "configurations": [
+         {
+             "name": "Run DesktopGL platform",
+             "type": "coreclr",
+             "request": "launch",
+             "preLaunchTask": "buildDesktopGL",
+             // If you have changed target frameworks, make sure to update the program path.
+             "program": "${workspaceFolder}/bin/Debug/net6.0/Tetris.dll",
+             "args": [],
+             "cwd": "${workspaceFolder}",
+             // For more information about the 'console' field, see https://github.com/OmniSharp/omnisharp-vscode/blob/master/debugger-launchjson.md#console-terminal-window
+             "console": "internalConsole",
+             "stopAtEntry": false
+         }
+     ]
+ }
+
